listener/requirements.txt
websocket-client==1.3.1
influxdb==5.3.1
influxdb-client==1.29.1
